Develop v

+++ src/main/java/com/whydigit/wms/dto/BuyerOrderDetailsDTO.java
@@ -11,9 +11,9 @@ public class BuyerOrderDetailsDTO {
 	private Long id;
 	private String partNo;
 	private String partDesc;
-	private Long qty;
+	private int qty;
 	private String batchNo;
-	private Long availQty;
+	private int availQty;
 	private String sku;
 	private String remarks;
 	private boolean qcflag;
